create CellGame component & main game board
